pyproject.toml
[project]
name = "kad-parser"
version = "0.1.0"
description = "Парсер документов из системы арбитражных судов РФ (КАД Арбитр)"
authors = [{name = "Andrew821667"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.8",
    "celery>=5.4.0",
    "httpx>=0.27.2",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "pdfplumber>=0.11.4",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.12",
    "jinja2>=3.1.4",
    "aiofiles>=24.1.0",
    "typer>=0.12.5",
    "rich>=13.8.1",
    "structlog>=24.4.0",
    "minio>=7.2.9",
    "playwright>=1.47.0",
    "curl-cffi>=0.7.1",
    "aiohttp-socks>=0.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "openpyxl>=3.1.5",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.2",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "types-beautifulsoup4>=4.12.0.20240907",
    "types-redis>=4.6.0.20240903",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "ARG", "PTH", "ERA", "PL", "RUF"]
ignore = ["E501", "PLR0913", "PLR2004"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[project.scripts]
kad-parser = "src.cli.commands:app"
